Extract keyboard addon
